Refactored RASCIL warning suppression. Fixed typo. :tea:
